+++ resources/scripts/commands/apexCodeCompletion.js
@@ -7,6 +7,7 @@ const applicationContext = require('../core/applicationContext');
 const NotificationManager = require('../output/notificationManager');
 const Editor = require('../output/editor');
 const TemplateUtils = require('../utils/templateUtils');
+const Config = require('../core/config');
 const window = vscode.window;
 const Range = vscode.Range;
 const SnippetString = vscode.SnippetString;
@@ -40,7 +41,7 @@ function processCommentCompletion(position, editor) {
     let declarationLine = editor.document.lineAt(lineNum);
     if (!declarationLine.isEmptyOrWhitespace) {
         const node = new ApexParser().setContent(FileReader.readDocument(editor.document)).setSystemData(applicationContext.parserData).setCursorPosition(position).isDeclarationOnly(true).parse();
-        const templateContent = TemplateUtils.getApexCommentTemplate();
+        const templateContent = TemplateUtils.getApexCommentTemplate(!Config.getConfig().documentation.useStandardJavaComments);
         if (templateContent) {
             applicationContext.parserData.template = templateContent;
             const apexComment = SnippetUtils.getApexComment(node, applicationContext.parserData.template, editor.document.uri.fsPath);